Moved view to correct folder and remove unused references
