+++ src/InventionDice/InventionDice.iOS/AppDelegate.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-
-using Foundation;
+﻿using Foundation;
 using UIKit;
 
 namespace InventionDice.iOS
