+++ src/InventionDice/InventionDice.iOS/Main.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-
-using Foundation;
-using UIKit;
+﻿using UIKit;
 
 namespace InventionDice.iOS
 {
