+++ src/InventionDice/InventionDice/Infrastructure/ApplicationConstants.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace InventionDice.Infrastructure
+﻿namespace InventionDice.Infrastructure
 {
     internal static class ApplicationConstants
     {
