+++ src/InventionDice/InventionDice/Services/Navigation/ViewModelFactory.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using InventionDice.ViewModels;
 
 namespace InventionDice.Services.Navigation
