+++ src/InventionDice/InventionDice/Views/MainPage.xaml
@@ -4,7 +4,7 @@
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d"
-             x:Class="InventionDice.MainPage">
+             x:Class="InventionDice.Views.MainPage">
 
     <StackLayout>
         <!-- Place new controls here -->
